Initial version of a docker based development environment
