+++ infra/rabbitmq/rabbitmq.conf
@@ -1,0 +1,12 @@
+listeners.ssl.default = 5671
+
+ssl_options.cacertfile = /etc/rabbitmq/certificates/testca/cacert.pem
+ssl_options.certfile   = /etc/rabbitmq/certificates/server/cert.pem
+ssl_options.keyfile    = /etc/rabbitmq/certificates/server/key.pem
+ssl_options.verify     = verify_peer
+ssl_options.fail_if_no_peer_cert = false
+
+auth_mechanisms.0 = PLAIN
+auth_mechanisms.1 = EXTERNAL
+
+load_definitions = /etc/rabbitmq/definitions.json
