+++ infra/tls/certificates/.gitignore
@@ -1,0 +1,1 @@
+*
