+++ infra/tools/pamqp-format
@@ -1,0 +1,8 @@
+#!/usr/bin/env sh
+
+set -o errexit
+set -o nounset
+
+base_dir=`dirname $0`/../../
+
+clang-format-17 -i $base_dir/*.c $base_dir/*.h
