+++ infra/tools/pamqp-stubs-validate
@@ -1,0 +1,19 @@
+#!/usr/bin/env sh
+
+set -o errexit
+set -o nounset
+
+SHARED_JSON=`php -n -r 'echo !extension_loaded("json") ? "true" : "false";'`
+
+args="-n"
+if [ "${SHARED_JSON}" = "true" ]; then
+  args="${args} -d extension=json.so"
+fi
+
+base_dir=`dirname $0`/../../
+real_base_dir=`realpath $base_dir`
+
+php $args $real_base_dir/infra/tools/pamqp-dump-reflection stubs.json
+php $args -d extension=modules/amqp.so $real_base_dir/infra/tools/pamqp-dump-reflection impl.json
+
+diff -10 -u stubs.json impl.json
