--- provision/nginx/default
@@ -1,21 +1,0 @@
-server {
-	listen 80 default_server;
-	listen [::]:80 default_server ipv6only=on;
-
-	root /var/www/html;
-
-	server_name _;
-
-    location / {
-        try_files $uri $uri/ /index.php?$args;
-        index index.php index-nginx.html;
-    }
-
-    location ~ \.php$ {
-        fastcgi_pass   127.0.0.1:9000;
-        fastcgi_index  index.php;
-        fastcgi_param  SCRIPT_FILENAME  $document_root$fastcgi_script_name;
-        include        fastcgi_params;
-        fastcgi_intercept_errors off;
-    }
-}
