--- provision/php/amqp.ini
@@ -1,3 +1,0 @@
-; configuration for php AMQP module
-; priority=10
-extension=amqp.so
