--- provision/test_certs/client/cert.pem
@@ -1,19 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDIzCCAgsCCQD/3A+eiHuqpzANBgkqhkiG9w0BAQsFADBFMQswCQYDVQQGEwJQ
-VDEPMA0GA1UECAwGTGlzYm9hMQ8wDQYDVQQHDAZMaXNib2ExFDASBgNVBAoMC0V4
-YW1wbGUgT3JnMB4XDTE4MDYxMDEwMzA0MFoXDTE5MDYxMDEwMzA0MFowYjELMAkG
-A1UEBhMCUFQxDzANBgNVBAgMBkxpc2JvYTEPMA0GA1UEBwwGTGlzYm9hMRQwEgYD
-VQQKDAtFeGFtcGxlIE9yZzEbMBkGA1UEAwwSY2xpZW50LmV4YW1wbGUub3JnMIIB
-IjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAzmgZMCuXgmc4moKvweny7mXT
-gKL8GNpB9uLwpmr0np4veWZh2b5WFFy6oma/4hS072PRXR1+miaEZHvlCgRopKp7
-Lm4CF45E8YamG0MjfTxWtTf+JAXjYibijl94GQLkXmMM9OBY080r01szP2RXxkfU
-3UMf0HjiDcj5YyJn2LVwkgqRx6JYcTNm/wRhk82vw6nclB6jeZ0OcJK/4lrNCaX/
-VfPTnyZoCXRteFGBiCWmf/Y2uvtjUqpwSAIDvNQaV6WrcmdTBt8szbigc1wR2WyX
-NR3WKy2aCPOPbbsCLDA/bilFnKUN0Ho53g3e8J0VUZv4G9WZKf2GrhHWax3lAQID
-AQABMA0GCSqGSIb3DQEBCwUAA4IBAQBY/0zo+2kz2ws1X0nRfz1J0cPvamlQj/jI
-C1G6h964nDjmcljmvybWlX/TCWDjtTS/bkEjUSZhAAqfF5Zt44g1WkwE03vzdPgF
-hN06NNbc5dQMYFa5cUnHbE3gkbN31qCf8T/IjzOBaqxDspwZyzvAXQwKFPke6TuA
-XjYinMlEVNL3TytEf5PhdDjvssrRRR/lKHdFF9ABTN8w/UWyMKrawuYrWynk/IdH
-F0i9GAPvULeZKHpYytwtwwitnucxsVJkkz0JVJcsrGRgwiy4zmIjSjATVD4kB+Dy
-MAcQTltJIAu/PSPFWvVAA73Wh87FdqGb2rWBKfer9MdzX4h7HnQO
------END CERTIFICATE-----
