--- provision/test_certs/client/key.pem
@@ -1,27 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEpAIBAAKCAQEAzmgZMCuXgmc4moKvweny7mXTgKL8GNpB9uLwpmr0np4veWZh
-2b5WFFy6oma/4hS072PRXR1+miaEZHvlCgRopKp7Lm4CF45E8YamG0MjfTxWtTf+
-JAXjYibijl94GQLkXmMM9OBY080r01szP2RXxkfU3UMf0HjiDcj5YyJn2LVwkgqR
-x6JYcTNm/wRhk82vw6nclB6jeZ0OcJK/4lrNCaX/VfPTnyZoCXRteFGBiCWmf/Y2
-uvtjUqpwSAIDvNQaV6WrcmdTBt8szbigc1wR2WyXNR3WKy2aCPOPbbsCLDA/bilF
-nKUN0Ho53g3e8J0VUZv4G9WZKf2GrhHWax3lAQIDAQABAoIBAFeW/1Fcvk+9Q9z1
-pmYQxGF8XpORFo/pVuLRDTRh4OrKz/ShiNahGeBMNXsjRTxIczu/TSJJmOcLf1Uw
-6lIk6s8t9VYPot2iJwmjjSu2gwHrMBR47WUWBpcpTy0BwHJEsuqG+AGWjrtarsVj
-R00YZNMkumUmzgJmue8FoSb7e7wPB8IPmlxAj0DheqnOVwfBCMJTCNCbmA2fLJen
-avLFzoURjlT6SlcQEv8cmz7jMskbwCTSlgCw5VtdYAQQLN4UnSfbMmInVNY2sSlM
-uVlKdU8UxhA1QoW9EUju9q0zI7wD0W8k6+jaRxLe8LBkwI0jh/nAcZtaKs5cQdxp
-/dDb8U0CgYEA7X8Uj2bjHaq7BNZY60BdDDMVwxxfSJhl99nhaACRNFk5JSSoZDCH
-GIB+DN+SiNH/Voh65Grnotf/i+bvG0TbjAVOrK35T7F2dvHZ8OeyR1EAl9EgW3z7
-c5Wjj+XwqhJL2fW01xcxOoR/Qs6Y7WsFaWM9PI+OqV16pVyR26eZ6ScCgYEA3nzs
-7eXs79oPmRd011+P9843j3hHETqHdH2GtZ1ZwycmhcJsjUo0G9tCOSFcxDjKiUpG
-Hkic4yueF/p0cd6ClSW9B8Hh8NX2B/oGxtWERB3dN7mtKwxPvniJ5FZe2U9np8ib
-zQynU/Z5nIRbK4Oone/bfmatQggKEmJM5iClCZcCgYEAhHSWu+/PBTG+QlFloDby
-biZjocDGJ4/Pdu8OdtmyIjcM1vld7RrfjbvEEVvzttkgBlvx5kj1TW4YOSef/V5m
-7+3Z1hblcBvyjR4PxfsDBCFaKlHZMQ1AlFNFuLCui7vOkA4oQLnPm+pfq5vb0LYa
-e498jLUoBK99ApLhSldsUIsCgYAFcgFKgUofRLMFCDHv44LeicmKVWNH5KrUCw50
-+3jq8dZbh7qoapyD4gXo6AwhMVKjsZGtxBEy7ipcY+DduJhxlg7eVbx17IatA02P
-KEKoeCKl7oFygEajXsfhMsv3fZ4H2T0W86xUvb+UORkSI5LI6snwegrbht7YPFet
-ejcP2QKBgQCv+TNpIFFMlhWwCbzp/I1sKM1QrQWI+Io+djxlyzxU4lFX2VX2xM9s
-k6maQT5VA7PHvQ+/aMTusVHVC6Rw2/b6rjej4zs0Z+pFXXR8Z2IJGBPMUZC9SKgi
-pUZjKSy0fXI2aR4QaNuZeYOYz6IR5TAhnCku8U+BAqaQ/uTtxj3J5w==
------END RSA PRIVATE KEY-----
