--- provision/test_certs/client/req.pem
@@ -1,17 +1,0 @@
------BEGIN CERTIFICATE REQUEST-----
-MIICpzCCAY8CAQAwYjELMAkGA1UEBhMCUFQxDzANBgNVBAgMBkxpc2JvYTEPMA0G
-A1UEBwwGTGlzYm9hMRQwEgYDVQQKDAtFeGFtcGxlIE9yZzEbMBkGA1UEAwwSY2xp
-ZW50LmV4YW1wbGUub3JnMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA
-zmgZMCuXgmc4moKvweny7mXTgKL8GNpB9uLwpmr0np4veWZh2b5WFFy6oma/4hS0
-72PRXR1+miaEZHvlCgRopKp7Lm4CF45E8YamG0MjfTxWtTf+JAXjYibijl94GQLk
-XmMM9OBY080r01szP2RXxkfU3UMf0HjiDcj5YyJn2LVwkgqRx6JYcTNm/wRhk82v
-w6nclB6jeZ0OcJK/4lrNCaX/VfPTnyZoCXRteFGBiCWmf/Y2uvtjUqpwSAIDvNQa
-V6WrcmdTBt8szbigc1wR2WyXNR3WKy2aCPOPbbsCLDA/bilFnKUN0Ho53g3e8J0V
-UZv4G9WZKf2GrhHWax3lAQIDAQABoAAwDQYJKoZIhvcNAQELBQADggEBAEgsC/yd
-GAoFCtNvwLsdbZBSWiFroImdD0lU01sY7UtXzEdosjYESZkXVz2b1sh2TRKleCq6
-Go99vW/AYqu2ietx6Np2bO7uNS0VwjWDaqCZbpw/fTPXPORF6sirqmoq03g15EGf
-1xM90dD1/EMuAiszTnhtprmr7j3fFIqN1x/9wHYVNEgUaNxCeevbX9LzHYzajzvu
-hkzT8YRxASSUex/1FfJOY+UYYwf46bKlB0tlmDuLXWKSrnBDJUTKwundLerdj/y5
-gZ6evfkxwvZOvr4JTYPcq2NwbhMzTkSnHOSK2AhzLsFVSA6CWTBTFHwQNDSblQdq
-FMjb5jvuyG83DTA=
------END CERTIFICATE REQUEST-----
