--- provision/test_certs/sasl-client/cert.pem
@@ -1,19 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDKDCCAhACCQDa/hd6atDdTzANBgkqhkiG9w0BAQsFADBFMQswCQYDVQQGEwJQ
-VDEPMA0GA1UECAwGTGlzYm9hMQ8wDQYDVQQHDAZMaXNib2ExFDASBgNVBAoMC0V4
-YW1wbGUgT3JnMB4XDTE4MDYxMDEwMzAzOVoXDTE5MDYxMDEwMzAzOVowZzELMAkG
-A1UEBhMCUFQxDzANBgNVBAgMBkxpc2JvYTEPMA0GA1UEBwwGTGlzYm9hMRQwEgYD
-VQQKDAtFeGFtcGxlIE9yZzEgMB4GA1UEAwwXc2FzbC1jbGllbnQuZXhhbXBsZS5v
-cmcwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQDD+p1IkyU1/E+A2B7T
-YIQV5+F/S4MTBI5VYx7WKzH6PgeUxNTBTyTPC6KzYtfTg+Yo6NRQ9QwwO1oSB+1e
-2HBi1avvUGHeV+6Gbou0odoTlu/MbBFNnqfsLzvtm9vzgh3kBgdeaNuD/EDKa9yU
-/MKCKoOK0fXsZzbMWxdWTHe5UYrhQFaX8QaM0+rH/payIGWEtdmA6L3oCPqRAKws
-svMADZT01duuS0/igVY7ZsbqMxFWwg9fstahY00mRZgbeAtcQEdHP/NIzEMH9/U3
-RRK3g2QmQX9EWPb8YLEro/prrHukL7XjyRhMMjtJf9h7beT/u9Vik2N0/KY1m9vA
-V293AgMBAAEwDQYJKoZIhvcNAQELBQADggEBAB2/5h/xW5CoE0/TU3sR64Vyp1yf
-og4jSnLISX3PFZU6I5U9m4JuH79tOItWyJmwgFLtk+sBThbh6qHNO6wvFlu0BMlW
-RuAAmvk2wohNNcbCQRYp4fSKYoWd/1swK72KBAXW4rqXR7mofA6uybtPMDvKH6UC
-XqPV532EcovNbncWFGCl32ryIeGUvEtQh43I+ds2G2YYslSqs/0B4ZZ5YHudG6PY
-075cez6N/4tTS0gi5RnbsaFvfoXuQ7nfIpXIcYPmSvKA0du1/qCMHZ4u2suHsrrl
-s5PZnDqiP7+UvCPWMUh9DvA2n2sjucPWtNW73zstmIAeBbp2+HW/oHCHwfg=
------END CERTIFICATE-----
