--- provision/test_certs/sasl-client/key.pem
@@ -1,27 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEowIBAAKCAQEAw/qdSJMlNfxPgNge02CEFefhf0uDEwSOVWMe1isx+j4HlMTU
-wU8kzwuis2LX04PmKOjUUPUMMDtaEgftXthwYtWr71Bh3lfuhm6LtKHaE5bvzGwR
-TZ6n7C877Zvb84Id5AYHXmjbg/xAymvclPzCgiqDitH17Gc2zFsXVkx3uVGK4UBW
-l/EGjNPqx/6WsiBlhLXZgOi96Aj6kQCsLLLzAA2U9NXbrktP4oFWO2bG6jMRVsIP
-X7LWoWNNJkWYG3gLXEBHRz/zSMxDB/f1N0USt4NkJkF/RFj2/GCxK6P6a6x7pC+1
-48kYTDI7SX/Ye23k/7vVYpNjdPymNZvbwFdvdwIDAQABAoIBAAZ32AcULC8685jK
-CUJqthFPBOdBo8LV7Kiuy86/wDeFIpTrZiY3KXzX4nfPNFZbzRXekVtWqx9QZimg
-L1ppBFSOvsz4NzKybkHrlhrRGLR6z9FMY8nPCbXG3VB7S2gfDEOBW27nQQ97HNxR
-cQdNzFGle4/yXA6AXZOjEgMcJPttwb7yJ/kOGCWN/yFtmt8TORIjnY25WZWCkdIz
-REVkfSBGlN5TISoX49UAnAxAxVLMlyw+6m2JHYuk7smrfhxcgrWNGm+UL5zCcK3M
-ruDXDxo6JXcJ/8Tj9TnsJ0Tj5HYzZiLTbwGttsrZaTqLH8TbWre7EvJOEN4k4vgX
-82I8VPECgYEA6JJvYnPvaaniT4r8gFgH746pjxndBt6t7gRMpQ8iE67+RidqAeNt
-4iXAd0XlELWCFS5Mb4zzgF8W9qdnhWPhCYA12Nz0wz26ivzRv02IVoVBcMefU7ka
-kuccfiDSfnJgxq1rTXnLFC4703GDwgi9MD8QhP/H2sxa3LVdCdZj1Q8CgYEA17iE
-n/dCrwlGB7gCaCwpuSS3d5LZJCz2SluTbLQkCyKzF8dirfmbwD4uOWXYokE/P/6N
-NyUOvjPNaT4YNgEDpeQcbDRAKHmbE4DAG+56lPy7ax9CU3E89PZzC4G+vh9sGJds
-ui73c2vvDvcR8jvgsJmjEUIgunYBZ2NCE9CoTxkCgYEAnZHo6TlHMb2p2N5/qKz7
-43AyrQOG3oqsKKC9FO+l1NQw7nLxvYnK/vc46RZv1dxD3/nYS5Ohvo2uTzqoRpWe
-+ALneKBrLIR8CIK10PjEsGnkJnb5GY5F1NXEVigCtSzlKHaCRqGH14pjeiRkcmfd
-VfzEQnfRfgoKXZJ7EJkyVj0CgYA38Z5TbehZRGpo7guRwIJBhYge9nJhs1dcCUu+
-USlXyfPwIsEwpR8DSBzsvcDks08X5Yfx2SZtpTmMJZJZzwGHMBU+6n4JASB6elVX
-6QzZPoHMUhr9UMyiKpfDUC6LV9LMvdhwoGWEBWXF18VLsMOgiPeFLkUWlivfCBrG
-0MIVYQKBgEnK29Wm6ukYxLV9j1RufN4ivTSny5cViyfWPB7GN9rWAeyFn9pkL6+q
-O0O08dPRJTAi9/7fQmJNi/+w0Ui3ty6dEPTVWBkxqd3LfRYRo+wOt+9tE4CI4PkC
-d63+MvyalKXyUY8at2+oWJ8lOJ6QyBtHsYQoNfzHUlR0JyS0AHln
------END RSA PRIVATE KEY-----
