--- provision/test_certs/sasl-client/req.pem
@@ -1,17 +1,0 @@
------BEGIN CERTIFICATE REQUEST-----
-MIICrDCCAZQCAQAwZzELMAkGA1UEBhMCUFQxDzANBgNVBAgMBkxpc2JvYTEPMA0G
-A1UEBwwGTGlzYm9hMRQwEgYDVQQKDAtFeGFtcGxlIE9yZzEgMB4GA1UEAwwXc2Fz
-bC1jbGllbnQuZXhhbXBsZS5vcmcwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEK
-AoIBAQDD+p1IkyU1/E+A2B7TYIQV5+F/S4MTBI5VYx7WKzH6PgeUxNTBTyTPC6Kz
-YtfTg+Yo6NRQ9QwwO1oSB+1e2HBi1avvUGHeV+6Gbou0odoTlu/MbBFNnqfsLzvt
-m9vzgh3kBgdeaNuD/EDKa9yU/MKCKoOK0fXsZzbMWxdWTHe5UYrhQFaX8QaM0+rH
-/payIGWEtdmA6L3oCPqRAKwssvMADZT01duuS0/igVY7ZsbqMxFWwg9fstahY00m
-RZgbeAtcQEdHP/NIzEMH9/U3RRK3g2QmQX9EWPb8YLEro/prrHukL7XjyRhMMjtJ
-f9h7beT/u9Vik2N0/KY1m9vAV293AgMBAAGgADANBgkqhkiG9w0BAQsFAAOCAQEA
-YToEWLnXTEHvvHFtnypKJUmKuM4iVCYZ91BnATI302PQNObg/DpNxslsaOMQ/AD8
-sHHr2DCnKZvofzCwSkSleMqsUn8Z6WGkI5jrW/umDXlfdgwWD4NqmiwYB5uebVYU
-cQje2nJJX7ERxnqSVMz1JjQLIbIe/eq8PJ8tdkLy2lzc6p3HrvNphnyb/hidRt5y
-ydeOjQx71/eur3EVKigXcMdoLI9P/Ilo7ju9H4FjGsW1eRDn9LY/YvaJfze/yMNz
-tGgwcZZHKdjTyg/WZeVgdYQJInR81FLD82AgNr3fsKB2QMNtT7mwB5bAx6zKCBhP
-VVXvUmqBv2i5lEVOn1l1gA==
------END CERTIFICATE REQUEST-----
