--- provision/test_certs/server/cert.pem
@@ -1,19 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDHjCCAgYCCQCvidFhSQh+AzANBgkqhkiG9w0BAQsFADBFMQswCQYDVQQGEwJQ
-VDEPMA0GA1UECAwGTGlzYm9hMQ8wDQYDVQQHDAZMaXNib2ExFDASBgNVBAoMC0V4
-YW1wbGUgT3JnMB4XDTE4MDYxMDEwMzAzOVoXDTE5MDYxMDEwMzAzOVowXTELMAkG
-A1UEBhMCUFQxDzANBgNVBAgMBkxpc2JvYTEPMA0GA1UEBwwGTGlzYm9hMRQwEgYD
-VQQKDAtFeGFtcGxlIE9yZzEWMBQGA1UEAwwNKi5leGFtcGxlLm9yZzCCASIwDQYJ
-KoZIhvcNAQEBBQADggEPADCCAQoCggEBALYkPyHJFpch8hZ8H3qlKjay737i4ic+
-g6V9urG1p+jmo1QnKzVV++4Pl26owPoe+T50SKLhhTFHJMEWeFB56Z9i58bhTalF
-vv6XF6W0rx62Dqd9joBn5ij/90EjCOKXhDV/ec9s4l3mjXV9EHa8H89q5lxW5RcZ
-KIcXJQgvWVkPAOCvGg9DnYnQLZhvKSwU3rfrdvkTVyoGiNjemQh42ca5JSF0Oi8B
-4siqdAgsa7uYsS2DzBTsHCDucDkSYSrWe+oS/vDRC86A2eZk4abVIcBlHOlxcwMT
-OKQ3oH3czOr3Ly8a8tTien3+5AfwZPjHxRDHeEU2Edyx9TnhO9wrdq8CAwEAATAN
-BgkqhkiG9w0BAQsFAAOCAQEAI1cqejtWM50BEUAzRXKPhrQFYqbStoOPMIFOQe41
-ytb7dSSh0qNhESoUSsCqNEGknMWdj8G7H0CxM+ti5TRVoKXjzrV+Ga6kyH6Mn9wL
-hrf84jl6EJFIlJkyz6/khLDH5HsCk0AAkxb9pr7HEjqWO3MrE9cdoHxOunb/qqme
-KIKtNNEE9HPuzfQHTQBKTi2QmcUj2AZn9kW1BbR3dye9mkGv19L3E3eWmzcYedN4
-OCxmAj01j8WN2cQ/fD3mNJUtr7eoN6hx51Z9s5+LQtULpKRDwbygxNXwEeVsqy17
-e6BUVOoUk+b1tTbUqtL/jM3XerGaHNdJzVZKO/K0Ywapzw==
------END CERTIFICATE-----
