--- provision/test_certs/server/key.pem
@@ -1,27 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEpQIBAAKCAQEAtiQ/IckWlyHyFnwfeqUqNrLvfuLiJz6DpX26sbWn6OajVCcr
-NVX77g+XbqjA+h75PnRIouGFMUckwRZ4UHnpn2LnxuFNqUW+/pcXpbSvHrYOp32O
-gGfmKP/3QSMI4peENX95z2ziXeaNdX0Qdrwfz2rmXFblFxkohxclCC9ZWQ8A4K8a
-D0OdidAtmG8pLBTet+t2+RNXKgaI2N6ZCHjZxrklIXQ6LwHiyKp0CCxru5ixLYPM
-FOwcIO5wORJhKtZ76hL+8NELzoDZ5mThptUhwGUc6XFzAxM4pDegfdzM6vcvLxry
-1OJ6ff7kB/Bk+MfFEMd4RTYR3LH1OeE73Ct2rwIDAQABAoIBAQCYs4q6KHxn+G2d
-rDI4AICDm0BugIhZnm+ALwh2NHWxLKuBwdTXwl6UFz495/zhtVYApjy/UZL43UxT
-FSOJVKwolCL8+abCY1bc2o9YvUKT3M4sXXc67+BmPzJryZOJZL3SnCzjWcXuM3wl
-I9TJtq50lorKn48w7kar+xtw4UFnB2KXRnM7TZiS7kGmRNKZx+VvH3praiBLcwke
-ZgBmfdZUIHDbvRHahqQ9TeXp02J83zY32yBlTnOlV6jkfpSh41fYCvOihKzkwv2v
-78FBI+y6f9Sn2sXC5bqyV1W1t9Y6eDFIUNn+Cg36yl2gNH6u5eG4Rw7zDJj0PIOj
-3n+hktoBAoGBAOKUEsorPvzzbdzHfh05oXOWxPqehJY5sjm9Td2Yx0nyQjkf33PT
-bXX+3MzJ1LaLER0EnBmJ7VPyjaguMf1NxaWrfykzRILtVbynn5Plk8IfDcvl+jFW
-oJdcg19L7Ey53VVdeCtTCqBmYrsUZF+s4/C6CpFZpj5F96KCOFVhYnWZAoGBAM3L
-AMEySk/sYZ+XzqNvt9uXNmULN3oqiUqz30/hBAcGKCzhnARr3clb+7WYbbd/12fr
-ZWwqwiHYcgKn11PCRAGnOxFRCr+N5ev0dSiNDBTKbvPKvKwj0jZqD8NL0bsmJlH9
-KY1wYUN7Ph1r7uCaSzm7yPH7FrQgGHD7fTLRLOuHAoGBAMAN4Utqks6YjJQqqPSe
-yAKRVtBUk3A+4ZpdouXlvW8poQaIxMCKA2uUJ5aSsfI1OMFPCf3/0qJoo7hdx9O9
-G36NOpEyuqz7/cOWlgYoADH4XpppisRNeHw0V8rsMRTsKvdT7itTDVvezWtkIpyS
-kPeYyqJfyjFY30npdVSx4mJpAoGASO3NCapBzYsOmZkZwK+hr28p9qr/8QpL9y5B
-UDSE0dxrtmQJf0OeqHIxhIxNLX9fndM88RVCEO58kNZcJ7Grmg5ij2Nx9Kpbtb+/
-GSzLAD8xDLJnJHXZVDFH/sTKwZDmeZ3G8PnHbjupqpGKaQwk5oPW1XJO/Gx3XOqy
-1qRsT9cCgYEAilHVQmFJWhWzFqvFV2wmt8KtxvyfDGU7HAP3417AQhQUv0ZvSZ1U
-snH5kFc+OlZs1+86QyvIgiRl2nInVyEzGHZqE0/pp77nsZQcNX0xQGBa9x+3bM1f
-6DB4+W9pC3oiU16mXytqDdmAQ93Vkb+WCxqDk/JgTl24ByxVsOgllY0=
------END RSA PRIVATE KEY-----
