--- provision/test_certs/server/req.pem
@@ -1,17 +1,0 @@
------BEGIN CERTIFICATE REQUEST-----
-MIICojCCAYoCAQAwXTELMAkGA1UEBhMCUFQxDzANBgNVBAgMBkxpc2JvYTEPMA0G
-A1UEBwwGTGlzYm9hMRQwEgYDVQQKDAtFeGFtcGxlIE9yZzEWMBQGA1UEAwwNKi5l
-eGFtcGxlLm9yZzCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBALYkPyHJ
-Fpch8hZ8H3qlKjay737i4ic+g6V9urG1p+jmo1QnKzVV++4Pl26owPoe+T50SKLh
-hTFHJMEWeFB56Z9i58bhTalFvv6XF6W0rx62Dqd9joBn5ij/90EjCOKXhDV/ec9s
-4l3mjXV9EHa8H89q5lxW5RcZKIcXJQgvWVkPAOCvGg9DnYnQLZhvKSwU3rfrdvkT
-VyoGiNjemQh42ca5JSF0Oi8B4siqdAgsa7uYsS2DzBTsHCDucDkSYSrWe+oS/vDR
-C86A2eZk4abVIcBlHOlxcwMTOKQ3oH3czOr3Ly8a8tTien3+5AfwZPjHxRDHeEU2
-Edyx9TnhO9wrdq8CAwEAAaAAMA0GCSqGSIb3DQEBCwUAA4IBAQCwq50gKVOAiT9U
-QFHC6pC8l4fexD7/pATf28XNa8ILwsETmNNAIbU+ESulWm921Nrxv1TjK+6qdvs6
-gZrnt6fyJ8mHt5fHqQoLUyLjWnlSoAAF5EFdCNtASlhOXVlhy5N2h+D5wsKMV6a6
-jNGQzAgyQNf6AAEEM2J47JfrYOgiEY/WBYK6O39VNhxyPIFbnkJIZIEI4lbHZCub
-xRWMNyDV4ngZGZmu9r6SzVjG1XDPc65Z3qg2LvzDZad7/fEpkLbxfZI9JcCXvMm4
-u6f6LUYgXDe2BXuXeF1Mo4Y8x7A6Wt5RuZibNq+Ls00D7P+ilvQjBy2F9laI2Sqw
-wHp0RISb
------END CERTIFICATE REQUEST-----
