--- provision/test_certs/testca/cacert.pem
@@ -1,21 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDXTCCAkWgAwIBAgIJALr5iv/qdSQuMA0GCSqGSIb3DQEBCwUAMEUxCzAJBgNV
-BAYTAlBUMQ8wDQYDVQQIDAZMaXNib2ExDzANBgNVBAcMBkxpc2JvYTEUMBIGA1UE
-CgwLRXhhbXBsZSBPcmcwHhcNMTgwNjEwMTAzMDM5WhcNMTkwNjEwMTAzMDM5WjBF
-MQswCQYDVQQGEwJQVDEPMA0GA1UECAwGTGlzYm9hMQ8wDQYDVQQHDAZMaXNib2Ex
-FDASBgNVBAoMC0V4YW1wbGUgT3JnMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIB
-CgKCAQEAsSAEBvCMGgZUqA+gIpviF3TigeVdMi3wu+9QmIviUcoIAXHr9zMxUyAP
-dxzEzJwhKkLXpLSt5uSUb07UUg32cLQdMmPbecuug41vdCsTz/6RaZlB9UKddYIO
-M4FQ31GElOfM3PsDIHjENQsarVTyt+lhIOXQdVCCkP0ebBa53pJwp0NOSl+ay+eq
-W7MQn56pFaD+AWd7SEAd2c/YV/aWU/sKHDIz/Ud/5ayFS4hZ6HU8lTqyeROzXYBL
-z85wcp+xTwUbvCJXTVZ3C6N+88bpi+9/1ORtSSOj7chIQ8ZdX+c58dCo1nJxw6fY
-HAwYSRsllSOIVVmtlL/oBLrYCM3+wwIDAQABo1AwTjAdBgNVHQ4EFgQUxBYlRH1g
-hokMTkwNbZaB4PRG7+swHwYDVR0jBBgwFoAUxBYlRH1ghokMTkwNbZaB4PRG7+sw
-DAYDVR0TBAUwAwEB/zANBgkqhkiG9w0BAQsFAAOCAQEAEn3kN7VPWnFu1sIveQII
-Sh2zQQYdgMS1zCnTVsZykhCliNBKwlUm6k/GEucibiggL56PESH1sUNwDK7bCR9k
-ngLAeA5YcNPyGLw0x0wcz0JH9rPqF+z5tHI1IeIEzilkWUim0MGf30B/jfn6+d3v
-gCk9OM7+bbTvZAvLhslHGKpgNwMEEAWzj3O4rvxp7M50VQ5XT3lvRXgeSUTeoepT
-Jq7WQOKVNPrWxGOHDqhnQz33yd7m1M0hge/W0IrlkwXY0/SDtuETv+xqAiUHYlNl
-s3INAWNFIMX/d7vBA4hsTujJMA9VzE+5jYQC+juXWsse61jW1eA+Fiu6m6lkXQE3
-Aw==
------END CERTIFICATE-----
