--- provision/test_certs/testca/private/cakey.pem
@@ -1,28 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQCxIAQG8IwaBlSo
-D6Aim+IXdOKB5V0yLfC771CYi+JRyggBcev3MzFTIA93HMTMnCEqQtektK3m5JRv
-TtRSDfZwtB0yY9t5y66DjW90KxPP/pFpmUH1Qp11gg4zgVDfUYSU58zc+wMgeMQ1
-CxqtVPK36WEg5dB1UIKQ/R5sFrneknCnQ05KX5rL56pbsxCfnqkVoP4BZ3tIQB3Z
-z9hX9pZT+wocMjP9R3/lrIVLiFnodTyVOrJ5E7NdgEvPznByn7FPBRu8IldNVncL
-o37zxumL73/U5G1JI6PtyEhDxl1f5znx0KjWcnHDp9gcDBhJGyWVI4hVWa2Uv+gE
-utgIzf7DAgMBAAECggEAFdOoK2LBIQPesscaKKmk9tcEgpVNIr3eiIo5GFNElGed
-DPCI73FzhpbQ0RDOf5hNzfNmt3k8ZNOOJxlDuMjXsZPAdVEw3pVm7/ch4OLeEieA
-q86i/iIzXSOltFsfKCXXfLID4Mdd30aChzbTFsD9GAS+X1tykMJXFJ7cfN1gEjks
-+6puuQ3PNpn3A8/jzcJn1b8eGfuxpJLclWbi2Fs2mLA4ZNnLXt+/IaTfJpSTteRD
-cPvfIVekIageihx+2BFfZGaM+BTAwTRb/TnWS+I7eRPnPLu4CL9/auxOoZPE8SRE
-lSczfo3QPpXNpgQw/VKnFP5BsRwKsPArUWX99kP3AQKBgQDdNf1JNBmCXSmfJeEp
-NIlv6oHDCcx65KoZRujQA4p5Z4fOJ8VZo8GAfiT6Au6xJKZxK8i34xgrbhcSsXIL
-ztL94gd/p90BJy+qyBOvNSb2ZjHv3WA1Au0lBGIAKt+xsLTZIpf5DavnnWTbOq74
-vxpcQnGp9ZR2oPpGINKz+AIx4wKBgQDM+x5iTEWUbO4MS4ljoCPhjEfPSHGtlUYZ
-nRAbPw4w1Wku3GQ5TF1cDDyVMOXq37QgElvJOVWCUKe6ICqotg0hTRp5wtBHIAex
-942rzpWSrY7kRAHZixPgFks5lTYXbn4lrYAtw2ifT8khRr3mTmupgjNeVK8ol+Gl
-ZA4zJWcVoQKBgQCYpNO1utzLZ0v8ROA9VcJs1s5Z7p4KQ1XZAWRPfdIE+lrs9+an
-d2dkKieL2ZAp7pdnyoXD1lHsThhfztBas4pGpLz16riHPioXV961bSEIr/ZPhVwI
-2I8Zlw+k+/DGJaB1oPhrZHK17ZcV7EWm4f/nn/XdCNg8j9KDp8ydYUrqbwKBgAjc
-Xh7fry5QWyX66fB9jq+EUeJa/lcDmeknifezjRh9UTOuYUK801hUSxYj6/xwOSgd
-wwv6x4B/nhCErXNnkz+6Roe27Sv17X869UvU/VA+4mtpqS7PPUe1jwDpO1Jd+2QD
-kQPpa49fcpFWroTTJQJJ15CfVocJsb5liduaJU+hAoGBANhM7PdXv0kfeVlEAqGW
-lXU6UXNYmEU0jmTcG+qr8lirr4xT8q8J1dPf3Eieu8q+9LdTI1wcKL3O9dtmZudR
-fdMPBpVCfmI6e9APwnDN/vMoYlfAUptkgfrBmBzZt0jP79p5v5of4ObG7QfqYPgr
-BWup9HG/UkSnk46Tc9z2V+6S
------END PRIVATE KEY-----
