--- tools/dev-build.sh
@@ -1,5 +1,0 @@
-#!/usr/bin/env sh
-set -e
-phpize
-CFLAGS="-D_FORTIFY_SOURCE=2 -fstack-protector-strong -Wall -Werror" ./configure
-make clean all
